video player added, confirm box added on logout and extra files removed

+++ src/features/auth/logout/logout.js
@@ -9,10 +9,12 @@ const Logout = () => {
   const navigate = useNavigate();
 
   const logoutHandler = () => {
-    userctx.setUserOff();
-    config.WSTOKEN = null;
-    localStorage.clear();
-    navigate("/");
+    if (window.confirm("Are you sure you want to logout?")) {
+      userctx.setUserOff({ userAuth: {}, userInfo: {} });
+      config.WSTOKEN = null;
+      localStorage.clear();
+      navigate("/");
+    }
   };
 
   return (
